all task done

+++ lib/views/widgets/mycourses/rating/actionRatingAppBar.dart
@@ -1,22 +1,34 @@
 import 'package:flutter/material.dart';
+import 'package:get/get.dart';
 import 'package:get/get_state_manager/src/simple/get_state.dart';
+import 'package:iconsax/iconsax.dart';
 import 'package:learn_nova/controller/myCourses/ratingController.dart';
+import 'package:learn_nova/core/constant/AppImages.dart';
+import 'package:learn_nova/core/constant/AppRoutes.dart';
+import 'package:learn_nova/core/function/confirmationAlert.dart';
 
 class ActionRatingAppBar extends StatelessWidget {
   const ActionRatingAppBar({super.key});
 
   @override
   Widget build(BuildContext context) {
+    // هنا ما نعمل put، فقط نجيب الكنترولر الموجود
     return GetBuilder<RatingController>(
       builder: (controller) {
         return controller.userRating.isNotEmpty
             ? PopupMenuButton<int>(
                 onSelected: (value) {
                   if (value == 0) {
-                    controller.userRating.clear();
-                    controller.update();
+                    Get.toNamed(AppRoutes.editRatingPage);
                   } else if (value == 1) {
-                    controller.deleteUserRating();
+                    confirmationAelrt(
+                        Appimages.delete,
+                        'Are you sure you want to delete your rating ??',
+                        'No, keep it',
+                        'Yes, delete',
+                        Colors.red, () {
+                      controller.deleteUserRating();
+                    });
                   }
                 },
                 icon: const Icon(Icons.more_vert),
@@ -29,20 +41,34 @@ class ActionRatingAppBar extends StatelessWidget {
                   PopupMenuItem(
                     value: 0,
                     child: Row(
-                      children: const [
-                        Icon(Icons.mode_edit_outline_rounded, size: 30),
+                      children: [
+                        Icon(Iconsax.edit),
                         SizedBox(width: 10),
-                        Text("Evaluation adjustment"),
+                        Text(
+                          "Edit Rating",
+                          style: Theme.of(context)
+                              .textTheme
+                              .titleLarge!
+                              .copyWith(
+                                  fontSize: 17, fontWeight: FontWeight.normal),
+                        ),
                       ],
                     ),
                   ),
                   PopupMenuItem(
                     value: 1,
                     child: Row(
-                      children: const [
-                        Icon(Icons.delete_rounded, color: Colors.red, size: 30),
+                      children: [
+                        Icon(Iconsax.trash),
                         SizedBox(width: 10),
-                        Text("Delete the rating"),
+                        Text(
+                          "Delete Rating",
+                          style: Theme.of(context)
+                              .textTheme
+                              .titleLarge!
+                              .copyWith(
+                                  fontSize: 17, fontWeight: FontWeight.normal),
+                        ),
                       ],
                     ),
                   ),
